requirements.txt
pinecone-client>=3.0.0
sentence-transformers==2.6.1
llama-cpp-python==0.2.77
fastapi==0.110.0
uvicorn==0.27.1
numpy
